Moved chunk script away from cnv calling wf.
